Move net loading to canvas import service.
